backend/requirements.txt
django
djangorestframework
djoser
django-filter
psycopg2-binary
gunicorn
python-dotenv
Pillow
sorl-thumbnail
django-debug-toolbar 